ci_vae/requirements.txt
torch>=1.8.0
torchvision>=0.9.0
pandas
matplotlib
scikit-learn
umap-learn
python-igraph
imageio
numpy
